samples/python/requirements.txt
deepgram-sdk>=3.0.0
pytest>=8.0.0
pytest-timeout>=2.0.0
pytest-asyncio>=0.23.0
aiofiles>=23.0.0
